requirements.txt
fredapi==0.5.2
notion-client>=2.3.0
python-dotenv>=1.0.0
requests>=2.28.0
